提交代码

+++ src/main/java/com/usedBooks/frontStage/shopCart/shopCartDetail/pojo/ShopCartDetail.java
@@ -19,7 +19,7 @@ public class ShopCartDetail {
 
     private Integer bookId;                 //书籍id
 
-    private Double purchaseQuantity;        //购买数量
+    private Integer  purchaseQuantity;        //购买数量
 
     private Integer sellerId;               //卖家id
 }
